+++ RELEASE-NOTES.md
@@ -1,4 +1,4 @@
 ## 0.1
 
 * #1 Support for provider methods in modules
-
+* #9 Module superclass with noop implementation of Module.configure(..)